temporary check-in to save current state
I started a rather large refactoring - at least it took more than I thought, so I created the refactored classes beneath and in an extra micro-project.
Next up is merging the new with the old code, so I want to have a state that compiles and runs with both implementation - this is it!